pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "openai>=1.70.0",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
